==> notebooks/David/02-Mask_Datasets.ipynb <==
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm

import re

import spacy

# ! python -m spacy download en_core_web_sm

nlp = spacy.load("en_core_web_sm")

#Create an EntityRuler for some of the names
nameRuler = nlp.add_pipe("entity_ruler")
patterns = [
    {"label": "PERSON", "pattern": [{"lower": 'delainey'}]},
    {"label": "PERSON", "pattern": [{"lower": 'jmf'}]},
    {"label": "PERSON", "pattern": [{"lower": 'dave'}]},
    {"label": "PERSON", "pattern": [{"lower": 'forney'}]},
    {"label": "PERSON", "pattern": [{"lower": 'lloyd'}]},
    {"label": "PERSON", "pattern": [{"lower": 'phillip'}]},
    {"label": "PERSON", "pattern": [{"lower": 'tj'}]},
    {"label": "ORG", "pattern": [{"lower": 'ercot'}]},
]
nameRuler.add_patterns(patterns)

ent_replacements = {
    'PERSON': "Steve",
    'ORG': "Apple",
    'GPE': "Cupertino",
}

def change_ents(doc, ent_replacements):
    # find all the ents
    regex_ent_replacements = {key: "" for key in ent_replacements}
    for ent in doc.ents:
        if ent.label_ in ent_replacements:
            text = re.sub("\(.*|\).*|\+.*", "", ent.text)
            regex_ent_replacements[ent.label_] = f"{regex_ent_replacements[ent.label_]}|{text}"
    regex_ent_replacements = {key: regex_ent_replacements[key][1:] for key in regex_ent_replacements}

    # replace all the ends
    new_text = doc.text
    for ent_label in regex_ent_replacements:
        if regex_ent_replacements[ent_label] != "":
            try:
                new_text = re.sub(regex_ent_replacements[ent_label], ent_replacements[ent_label], new_text)
            except:
                continue
    new_text = re.sub("[ \n\t]+", " ", new_text)
    
    return new_text

email_file = "normal_emails.csv"

df = pd.read_csv(email_file, index_col='Original Index')
# token_list = [token for token in doc if " " not in token.text]

replaced_emails = [change_ents(doc, ent_replacements) for doc in nlp.pipe(tqdm(df['Email']))]

df['Classify Email'] = replaced_emails

df.to_csv(email_file)

email_file = "exec_emails.csv"

df = pd.read_csv(f"../../data/{email_file}", index_col='Original Index')
# token_list = [token for token in doc if " " not in token.text]

replaced_emails = [change_ents(doc, ent_replacements) for doc in nlp.pipe(tqdm(df['Email']))]

df['Classify Email'] = replaced_emails

df.to_csv(email_file)

email_file = "poi_emails.csv"

df = pd.read_csv(f"../../data/{email_file}", index_col='Original Index')
# token_list = [token for token in doc if " " not in token.text]

replaced_emails = [change_ents(doc, ent_replacements) for doc in nlp.pipe(tqdm(df['Email']))]

df['Classify Email'] = replaced_emails

df.to_csv(email_file)

replaced_emails = [change_ents(doc, ent_replacements) for doc in nlp.pipe(df['Email'])]

df['Classify Email'] = replaced_emails

df

# name = 'Skilling'
# for i in df[df['Sender'] == name]['Classify Email'].sample(10):
#     print(i)
#     print()

df.to_csv(email_file)
