==> notebooks/David/03-Dataset_Splitting.ipynb <==
## Dataset Splitting

import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split

import os
import re
import sys
import string
import pyprojroot

proj_root = pyprojroot.find_root(pyprojroot.has_file(".git"))
sys.path.append(os.path.join(proj_root, 'code'))

### 1) Split and prepare datasets

df_poi = pd.read_csv("poi_emails.csv", index_col='Original Index')
df_exec = pd.read_csv("exec_emails.csv", index_col='Original Index')
df_norm = pd.read_csv("normal_emails.csv", index_col='Original Index')

df_norm.POI.any()

def get_email_samples(df, email_cap=4000, max_sample_size=700):
    df_senders_list = []
    total_emails = 0
    for sender in df['Sender'].unique():
        df_tmp = df[df['Sender'] == sender]
        sample_size = np.min([max_sample_size, df_tmp.shape[0]])
        df_tmp = df_tmp.sample(sample_size)
        df_senders_list.append(df_tmp)
        total_emails += sample_size
        if total_emails > email_cap:
            break
    return pd.concat(df_senders_list)

email_sample_size = 700
num_exec_emails = 3000
num_norm_emails = 9000
df_exec_samples = get_email_samples(df_exec, email_cap=num_exec_emails, max_sample_size=email_sample_size)
df_norms_samples = get_email_samples(df_norm, email_cap=num_norm_emails, max_sample_size=email_sample_size)

df = pd.concat([df_poi, df_exec_samples, df_norms_samples]).sort_index()

num_emails = df.shape[0]
num_poi = df[df['POI']].shape[0]
num_execs = df[(~df['POI']) & df['Exec 200']].shape[0]
num_norms = df[(~df['POI']) & (~df['Exec 200'])].shape[0]

print(f"Number of Emails = {num_emails}")
print(f"Fraction of POI = {num_poi/num_emails}")
print(f"Fraction of Execs = {num_execs/num_emails}")
print(f"Fraction of POI = {num_norms/num_emails}")

# breakdown of senders
df[['Sender', 'POI', 'Exec 200']].value_counts().sort_index(level=["POI", "Exec 200"])

valid_size = 1986
test_size = 3310

# get Datetime objects
def date_standardizer(s):
    s = s.strip()
    m = re.search("[\d]+", s)
    if (m.end() - m.start()) == 1:
        return re.sub(f" {m.group()} ", f" 0{m.group()} ", s)
    return s
    
df['Date'] = df['Date'].apply(lambda x: date_standardizer(x))
df['Datetime'] = pd.to_datetime(df['Date'], format="%a, %d %b %Y %H:%M:%S")

df

y_full = df['POI']
X_full = df[[col for col in df.columns if col != 'POI']]

X_train, X_test, y_train, y_test = train_test_split(X_full, y_full, test_size=test_size, stratify=y)

X_train, X_valid, y_train, y_valid = train_test_split(X_train, y_train, test_size=valid_size, stratify=y_train)

print(f"Training: {X_train.shape[0]}, Validation: {X_valid.shape[0]}, Test: {X_test.shape[0]}")

X_test[['Sender', 'Exec 200']].value_counts().sort_index(level=["Exec 200"])

X_valid[['Sender', 'Exec 200']].value_counts().sort_index(level=["Exec 200"])

X_train[['Sender', 'Exec 200']].value_counts().sort_index(level=["Exec 200"])

X_train['POI'] = y_train
X_valid['POI'] = y_valid
X_test['POI'] = y_test

X_train[[col for col in X_train.columns if col != 'Datetime']].to_csv("train_set.csv")
X_valid[[col for col in X_valid.columns if col != 'Datetime']].to_csv("valid_set.csv")
X_test[[col for col in X_test.columns if col != 'Datetime']].to_csv("test_set.csv")
