==> notebooks/David/04-Modelling.ipynb <==
## Modeling

import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.metrics import precision_recall_fscore_support

import re
import string

df_train = pd.read_csv("train_set.csv", index_col='Original Index')
df_valid = pd.read_csv("valid_set.csv", index_col='Original Index')
df_test = pd.read_csv("test_set.csv", index_col='Original Index')

X_train = df_train[[col for col in df_train.columns if col != 'POI']]
y_train = df_train['POI']

X_valid = df_valid[[col for col in df_valid.columns if col != 'POI']]
y_valid = df_valid['POI']

X_test = df_test[[col for col in df_test.columns if col != 'POI']]
y_test = df_test['POI']

print(f"Training: {X_train.shape[0]}, Validation: {X_valid.shape[0]}, Test: {X_test.shape[0]}")

import nltk
from nltk.tokenize import word_tokenize

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

# get NLTK's stopwords
nltk.download('stopwords')
stopwords = set(nltk.corpus.stopwords.words("english"))

# POS tagging
from nltk.tag import pos_tag
nltk.download('averaged_perceptron_tagger')

# lemmatization
from nltk.stem import WordNetLemmatizer
from nltk.corpus import wordnet
 
lemmatizer = WordNetLemmatizer()

def pos_normalizer(nltk_tag):
    if nltk_tag.startswith('J'):
        return wordnet.ADJ
    elif nltk_tag.startswith('V'):
        return wordnet.VERB
    elif nltk_tag.startswith('N'):
        return wordnet.NOUN
    elif nltk_tag.startswith('R'):
        return wordnet.ADV   
    return wordnet.NOUN
  
# stemming
from nltk.stem.porter import PorterStemmer
stemmer = PorterStemmer()

# n-gram generator
from nltk import ngrams

# lemmatize, unigram
def preprocess_1(s):
    # tokenize
    tokens = [token for token in word_tokenize(s)]
    
    # remove numbers and apostrophes and convert to lowercase
    tokens = [token.lower() for token in tokens if len(re.findall("[“”'’`\d]+", token)) == 0]

    # assign POS
    pos_tags = pos_tag(tokens)

    # lemmatize
    lemmas = [lemmatizer.lemmatize(token, pos=pos_normalizer(pos)) for token, pos in pos_tags]

    # remove stopwords and punctuation
    tokens = [token for token in lemmas if (token not in stopwords and token not in string.punctuation)]

    # n-gram
    n = 1
    feats = [" ".join(gram) for gram in list(ngrams(tokens, n))]
    
    return feats

# stem, unigram
def preprocess_2(s):
    # tokenize
    tokens = [token for token in word_tokenize(s)]
    
    # remove numbers and apostrophes and convert to lowercase
    tokens = [token.lower() for token in tokens if len(re.findall("[“”'’`\d]+", token)) == 0]

    # assign POS
    pos_tags = pos_tag(tokens)
 
    # stem
    stems = [stemmer.stem(word) for word in tokens]

    # remove stopwords and punctuation
    tokens = [token for token in stems if (token not in stopwords and token not in string.punctuation)]

    # n-gram
    n = 1
    feats = [" ".join(gram) for gram in list(ngrams(tokens, n))]
    
    return feats

# lemmatize, bigram
def preprocess_3(s):
    # tokenize
    tokens = [token for token in word_tokenize(s)]
    
    # remove numbers and apostrophes and convert to lowercase
    tokens = [token.lower() for token in tokens if len(re.findall("[“”'’`\d]+", token)) == 0]

    # assign POS
    pos_tags = pos_tag(tokens)

    # lemmatize
    lemmas = [lemmatizer.lemmatize(token, pos=pos_normalizer(pos)) for token, pos in pos_tags]

    # remove stopwords and punctuation
    tokens = [token for token in lemmas if (token not in stopwords and token not in string.punctuation)]

    # n-gram
    n = 2
    feats = [" ".join(gram) for gram in list(ngrams(tokens, n))]
    
    return feats

# stem, bigram
def preprocess_4(s):
    # tokenize
    tokens = [token for token in word_tokenize(s)]
    
    # remove numbers and apostrophes and convert to lowercase
    tokens = [token.lower() for token in tokens if len(re.findall("[“”'’`\d]+", token)) == 0]

    # assign POS
    pos_tags = pos_tag(tokens)
 
    # stem
    stems = [stemmer.stem(word) for word in tokens]

    # remove stopwords and punctuation
    tokens = [token for token in stems if (token not in stopwords and token not in string.punctuation)]

    # n-gram
    n = 2
    feats = [" ".join(gram) for gram in list(ngrams(tokens, n))]
    
    return feats

# # lemmatize, unigram, no stopword removal
# def preprocess_5(s):
#     # tokenize
#     tokens = [token for token in word_tokenize(s)]
    
#     # remove numbers and apostrophes and convert to lowercase
#     tokens = [token.lower() for token in tokens if len(re.findall("[“”'’`\d]+", token)) == 0]

#     # assign POS
#     pos_tags = pos_tag(tokens)

#     # lemmatize
#     lemmas = [lemmatizer.lemmatize(token, pos=pos_normalizer(pos)) for token, pos in pos_tags]

#     # remove punctuation
#     tokens = [token for token in lemmas if (token not in string.punctuation)]

#     # n-gram
#     n = 1
#     feats = [" ".join(gram) for gram in list(ngrams(tokens, n))]
    
#     return feats

# # stem, unigram, no stopword removal
# def preprocess_6(s):
#     # tokenize
#     tokens = [token for token in word_tokenize(s)]
    
#     # remove numbers and apostrophes and convert to lowercase
#     tokens = [token.lower() for token in tokens if len(re.findall("[“”'’`\d]+", token)) == 0]

#     # assign POS
#     pos_tags = pos_tag(tokens)
 
#     # stem
#     stems = [stemmer.stem(word) for word in tokens]

#     # remove punctuation
#     tokens = [token for token in stems if (token not in string.punctuation)]

#     # n-gram
#     n = 1
#     feats = [" ".join(gram) for gram in list(ngrams(tokens, n))]
    
#     return feats

# # lemmatize, bigram, no stopword removal
# def preprocess_7(s):
#     # tokenize
#     tokens = [token for token in word_tokenize(s)]
    
#     # remove numbers and apostrophes and convert to lowercase
#     tokens = [token.lower() for token in tokens if len(re.findall("[“”'’`\d]+", token)) == 0]

#     # assign POS
#     pos_tags = pos_tag(tokens)

#     # lemmatize
#     lemmas = [lemmatizer.lemmatize(token, pos=pos_normalizer(pos)) for token, pos in pos_tags]

#     # remove punctuation
#     tokens = [token for token in lemmas if (token not in string.punctuation)]

#     # n-gram
#     n = 2
#     feats = [" ".join(gram) for gram in list(ngrams(tokens, n))]
    
#     return feats

# # stem, bigram, no stopword removal
# def preprocess_8(s):
#     # tokenize
#     tokens = [token for token in word_tokenize(s)]
    
#     # remove numbers and apostrophes and convert to lowercase
#     tokens = [token.lower() for token in tokens if len(re.findall("[“”'’`\d]+", token)) == 0]

#     # assign POS
#     pos_tags = pos_tag(tokens)
 
#     # stem
#     stems = [stemmer.stem(word) for word in tokens]

#     # remove punctuation
#     tokens = [token for token in stems if (token not in string.punctuation)]

#     # n-gram
#     n = 2
#     feats = [" ".join(gram) for gram in list(ngrams(tokens, n))]
    
#     return feats

from sklearn.naive_bayes import MultinomialNB, ComplementNB
from sklearn.linear_model import LogisticRegression

def train_model(X_train, y_train, preprocess_func=preprocess_1, model_type="naive", hyperparam=1):
    
    # preprocess the training set
    X_processed = [preprocess_func(x) for x in X_train]
    
    # initialize vectorizer
    vectorizer = TfidfVectorizer(analyzer=(lambda x: x))
        
    # fit vectorizer
    X = vectorizer.fit_transform(X_processed)
    
    # initialize the classifier
    if model_type == "naive":
        # clf = MultinomialNB(alpha=hyperparam)
        clf = ComplementNB(alpha=hyperparam)
    elif model_type == "svm":
        clf = SVC(C=hyperparam)
    elif model_type == "logistic":
        clf = LogisticRegression(C=hyperparam)
    
    # fit the model
    clf.fit(X, y_train)
    
    return clf, vectorizer

def get_preds(X, clf, vectorizer, preprocess_func):
    X_processed = [preprocess_func(x) for x in X]
    X_vectorized = vectorizer.transform(X_processed)
    return clf.predict(X_vectorized)

def compute_f1(X, y_true, clf, vectorizer, preprocess_func):
    y_pred = get_preds(X, clf, vectorizer, preprocess_func)
    return precision_recall_fscore_support(y_true, y_pred)[2]

def compute_acc(X, y_true, clf, vectorizer, preprocess_func):
    X_processed = [preprocess_func(x) for x in X]
    X_vectorized = vectorizer.transform(X_processed)
    return clf.score(X_vectorized, y_true)

email_col = 'Classify Email'

preprocessors = [preprocess_1, preprocess_2, preprocess_3, preprocess_4]
model_types = ["naive", "logistic"]
hyperparams = [0.5, 1, 10]
# model_types = ["logistic"]
# hyperparams = [10, 20, 50, 100]

preprocessing_dict = {
    preprocess_1 : "Lemmatize + Unigram",
    preprocess_2 : "Stem + Unigram",
    preprocess_3 : "Lemmatize + Bigram",
    preprocess_4 : "Stem + Bigram"
}

best_model = {
    "preprocessor": preprocess_1,
    "model": "naive",
    "hyperparam": 0.5,
    "valid_acc": 0
}

for preprocessor in preprocessors:
    print(f"Preprocessor: {preprocessing_dict[preprocessor]}")
    for model_type in model_types:
        print(f"Model: {model_type}")
        for hyperparam in hyperparams:
            if model_type == "naive":
                print(f"alpha = {hyperparam}")
            else:
                print(f"C = {hyperparam}")
                
            clf, vectorizer = train_model(X_train[email_col], y_train, preprocess_func=preprocessor, model_type=model_type, hyperparam=hyperparam)
            
            valid_f1 = compute_f1(X_valid[email_col], y_valid, clf, vectorizer, preprocessor)
            valid_acc = compute_acc(X_valid[email_col], y_valid, clf, vectorizer, preprocessor)

            print(f"\tValidation f1: {np.round(valid_f1[1], 3)*100}") 
            print(f"\tValidation accuracy: {np.round(valid_acc, 3)*100}") 
            
            # save model, if it's the best
            if valid_acc > best_model['valid_acc']:
                best_model["preprocessor"] = preprocessor
                best_model["model"] = model_type
                best_model["hyperparam"] = hyperparam
                best_model["valid_acc"] = valid_acc
        print()
    print()

print("Best model:")
print(f"\tPreprocessing: {preprocessing_dict[best_model['preprocessor']]}")
print(f"\tModel: {best_model['model']}")
if best_model['model'] == 'naive':
    print(f"\tHyperparameter: alpha = {best_model['hyperparam']}")
else:
    print(f"\tHyperparameter: C = {best_model['hyperparam']}")

clf, vectorizer = train_model(X_train[email_col], y_train, preprocess_func=best_model["preprocessor"], model_type=best_model["model"], hyperparam=best_model["hyperparam"])
test_f1 = compute_f1(X_test[email_col], y_test, clf, vectorizer, best_model['preprocessor'])
test_acc = compute_acc(X_test[email_col], y_test, clf, vectorizer, best_model['preprocessor'])
print(f"Final test f1: {test_f1[1]}")
print(f"Final test accuracy: {test_acc}")

# preprocess_func = best_model["preprocessor"]
# model_type = best_model["model"]
# hyperparam = best_model["hyperparam"]

# # preprocess the training set
# X_processed = [preprocess_func(x) for x in X_train[email_col]]

# # initialize vectorizer
# vectorizer = TfidfVectorizer(analyzer=(lambda x: x))

# # fit vectorizer
# X = vectorizer.fit_transform(X_processed)

# # initialize the classifier
# if model_type == "naive":
#     # clf = MultinomialNB(alpha=hyperparam)
#     clf = ComplementNB(alpha=hyperparam)
# elif model_type == "svm":
#     clf = SVC(C=hyperparam)
# elif model_type == "logistic":
#     clf = LogisticRegression(C=hyperparam)

# # fit the model
# clf.fit(X, y_train)

# # evaluate on the validation set
# X_valid_processed = [preprocess_func(x) for x in X_valid[email_col]]
# X_v = vectorizer.transform(X_valid_processed)

# # evaluate on the test set
# X_test_processed = [preprocess_func(x) for x in X_test[email_col]]
# X_t = vectorizer.transform(X_test_processed)

# y_pred = clf.predict(X_t)

y_pred = get_preds(X_test[email_col], clf, vectorizer, best_model['preprocessor'])

inv_vocab = {vectorizer.vocabulary_[key]:key for key in vectorizer.vocabulary_}

coefs = pd.DataFrame(clf.coef_.reshape((-1,)), columns=["Parameter"])

coefs['Parameter Positive'] = coefs['Parameter'].apply(lambda x: 0 if x < 0 else 1)
coefs["|Parameter|"] = np.abs(coefs['Parameter'])

coefs['Word'] = [inv_vocab[key] for key in coefs.index]

coefs.sort_values('|Parameter|', ascending=False).head(20)

prec, rec, f1, supp = precision_recall_fscore_support(y_test, y_pred)

print(f"Precision:\n\tNon-POI = {np.round(prec[0], 4)}\n\tPOI = {np.round(prec[1], 4)}")
print(f"Recall:\n\tNon-POI = {np.round(rec[0], 4)}\n\tPOI = {np.round(rec[1], 4)}")
print(f"F1:\n\tNon-POI = {np.round(f1[0], 4)}\n\tPOI = {np.round(f1[1], 4)}")
print(f"Support:\n\tNon-POI = {np.round(supp[0], 4)}\n\tPOI = {np.round(supp[1], 4)}")

# Lots of false negatives -> we're missing a lot of positives

# poi_idx = y_test
# exec_idx = (X_test['Exec 300']) & (~y_test)
# norm_idx = (~X_test['Exec 300']) & (~y_test)

# y_pred_poi = get_preds(X_test.loc[poi_idx, email_col], clf, vectorizer, best_model['preprocessor'])
# y_pred_exec = get_preds(X_test.loc[exec_idx, email_col], clf, vectorizer, best_model['preprocessor'])
# y_pred_norm = get_preds(X_test.loc[norm_idx, email_col], clf, vectorizer, best_model['preprocessor'])

# y_true_poi = y_test[poi_idx]
# y_true_exec = y_test[exec_idx]
# y_true_norm = y_test[norm_idx]

# k = {
#     "POI": (y_true_poi, y_pred_poi), 
#     "Exec": (y_true_exec, y_pred_exec), 
#     "Normal": (y_true_norm, y_pred_norm) 
# }

# for type_ in k:
#     print(type_)
#     y_t, y_p = k[type_]
#     prec, rec, f1, supp = precision_recall_fscore_support(y_t, y_p)
    
#     print(f"Precision:\n\tNon-POI = {np.round(prec[0], 4)}\n\tPOI = {np.round(prec[1], 4)}")
#     print(f"Recall:\n\tNon-POI = {np.round(rec[0], 4)}\n\tPOI = {np.round(rec[1], 4)}")
#     print(f"F1:\n\tNon-POI = {np.round(f1[0], 4)}\n\tPOI = {np.round(f1[1], 4)}")
#     print(f"Support:\n\tNon-POI = {np.round(supp[0], 4)}\n\tPOI = {np.round(supp[1], 4)}")
#     print()
